add channel class and modify node for multi topic
